Config file bug cleanup (#370)
Fixed a couple of bugs regarding configuration options. The two fixed in this PR are described as follows:

[FIXED] Vector-Length being greater than Load-Bandwidth or Store-Bandwidth caused progression to halt with no error. A simple check is added in ModelConfig.cc to ensure this is not the case, and to error out if it is the case. Any tests impacted by this are fixed

[FIXED] Fetch-Block-Size being greater than 128 (despite being a valid value) caused a large memory leak. This was due to the precision of bufferedBytes_ and bufferOffset being 8 bit, as was the bytesAvailable parameter of the function predecode. This led to an infinite loop of instruction fetches. This has been solved by changing the precision to 16 bit, which covers the full range of configuration options for Fetch-Block-Size.

+++ src/include/simeng/arch/Architecture.hh
@@ -65,7 +65,7 @@ class Architecture {
    * Writes into the supplied macro-op vector, and returns the number of bytes
    * consumed to produce it; a value of 0 indicates too few bytes were present
    * for a valid decoding. */
-  virtual uint8_t predecode(const void* ptr, uint8_t bytesAvailable,
+  virtual uint8_t predecode(const void* ptr, uint16_t bytesAvailable,
                             uint64_t instructionAddress,
                             MacroOp& output) const = 0;
